chore : chatroomservice 위치 이동
- chatRoomService 객체를 application의 service 폴더로 이동
