+++ chat-service/src/main/java/com/sparta/moim/chat/presentation/controller/ChatRoomController.java
@@ -4,17 +4,14 @@ import static com.sparta.moim.chat.infrastructure.response.ChatCode.*;
 
 import com.sparta.moim.chat.application.dto.ChatRoomResponseDTO;
 import com.sparta.moim.chat.application.service.ChatRoomService;
-import com.sparta.moim.chat.domain.service.ChatRoomDomainService;
 import com.sparta.moim.chat.infrastructure.response.ChatCode;
 import com.sparta.moim.chat.presentation.request.ChatRoomRequestDTO;
 import com.sparta.moim.common.response.ApiResponseData;
-import com.sparta.moim.common.response.CommonCode;
 import com.sparta.moim.common.security.CustomUserDetails;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -31,12 +28,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class ChatRoomController {
 
   private final ChatRoomService chatRoomService;
-  private final ChatRoomDomainService chatRoomDomainService;
 
   // 채팅방 생성하기, 관리자만 가능
   @PostMapping("")
   public ResponseEntity<ApiResponseData<ChatRoomResponseDTO>> createChatRoom(@RequestBody ChatRoomRequestDTO chatRoomRequestDTO){
-
     return ResponseEntity.ok().body(ApiResponseData.of(CHAT_ROOM_FOUND.getCode() , CHAT_ROOM_FOUND.getMessage(), chatRoomService.createChatRoomService(chatRoomRequestDTO)));
   }
 
@@ -44,23 +39,24 @@ public class ChatRoomController {
   // 기획상 채팅방은 현재 1개 이지만, 확장성을 고려하여 채팅방 여러개 조회할 수 있도록 구현함
   @GetMapping("/{organizationId}")
   public ResponseEntity<ApiResponseData<List<ChatRoomResponseDTO>>> readChatRoom(@PathVariable("organizationId") String organizationId){
-    return ResponseEntity.ok().body(ApiResponseData.of(ChatCode.CHAT_ROOM_FOUND.getCode(), ChatCode.CHAT_ROOM_FOUND.getMessage(),chatRoomDomainService.readChatRooms(organizationId)));
+    return ResponseEntity.ok().body(ApiResponseData.of(ChatCode.CHAT_ROOM_FOUND.getCode(), ChatCode.CHAT_ROOM_FOUND.getMessage(),chatRoomService.readChatRooms(organizationId)));
   }
 
   // 채팅방 수정하기 - 채팅방 이름만 수정 가능함
   @PutMapping("/{chatRoomId}")
   public ResponseEntity<ApiResponseData<String>> updateChatRoom(@PathVariable("chatRoomId") String chatRoomId,
                                                                 @RequestBody @Valid ChatRoomRequestDTO chatRoomRequestDTO){
-    chatRoomDomainService.updateChatRoom(chatRoomId,chatRoomRequestDTO);
+    chatRoomService.updateChatRoom(chatRoomId,chatRoomRequestDTO);
     return ResponseEntity.ok().body(ApiResponseData.of(CHAT_ROOM_UPDATE.getCode(), CHAT_ROOM_UPDATE.getMessage(),null));
   }
 
-  // 채팅방 삭제하기
+  // 채팅방 접속 끊기 (채팅방 삭제하기)
   @DeleteMapping("/{chatRoomId}")
   public ResponseEntity<ApiResponseData<String>> deleteChatRoom(@PathVariable("chatRoomId") String chatRoomId,
                                                                 @AuthenticationPrincipal CustomUserDetails customUserDetails){
-    chatRoomDomainService.deleteChatRoom(chatRoomId, customUserDetails);
+    chatRoomService.deleteChatRoom(chatRoomId, customUserDetails);
     return ResponseEntity.ok().body(ApiResponseData.of(CHAT_ROOM_DELETE.getCode(), CHAT_ROOM_DELETE.getMessage(),null));
   }
+
 
 }
